requirements.txt
amqp==2.5.2
appnope==0.1.0
aws-xray-sdk==0.95
backcall==0.1.0
billiard==3.5.0.5
boto==2.49.0
boto3==1.9.253
botocore==1.12.253
celery==4.2.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
defusedxml==0.6.0
Django==2.1.15
django-appconf==1.0.3
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-cors-headers==2.4.1
django-crontab==0.7.1
django-extensions==2.2.5
django-filter==2.2.0
django-imagekit==4.0.2
django-nested-admin==3.1.3
django-storages==1.7.2
django-timezone-field==4.0
djangorestframework==3.9.4
docker==4.1.0
docutils==0.15.2
ecdsa==0.14.1
future==0.18.2
idna==2.7
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
Jinja2==2.10.3
jmespath==0.9.4
jsondiff==1.1.1
jsonpickle==1.2
kombu==4.3.0
MarkupSafe==1.1.1
mixins==0.1.1
mock==3.0.5
moto==1.3.7
oauthlib==3.1.0
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
pilkit==2.0
Pillow==5.3.0
prompt-toolkit==3.0.2
psycopg2-binary==2.7.7
ptyprocess==0.6.0
pyaml==19.12.0
pycparser==2.19
pycryptodome==3.9.4
Pygments==2.5.2
python-crontab==2.4.0
python-dateutil==2.8.1
python-jose==2.0.2
python-monkey-business==1.0.0
python3-openid==3.1.0
pytz==2018.7
PyYAML==5.2
redis==3.2.1
requests==2.20.1
requests-oauthlib==1.3.0
responses==0.10.9
s3transfer==0.2.1
selenium==3.141.0
six==1.13.0
traitlets==4.3.3
urllib3==1.24.3
vine==1.3.0
wcwidth==0.1.8
websocket-client==0.57.0
Werkzeug==0.16.0
wrapt==1.11.2
xmltodict==0.12.0
